+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Changed
+
+- upgraded the Toolbox plugin API, dropping support for Toolbox versions older than 3.7.2
+
 ## 0.9.4 - 2026-08-26
 
 ### Added